move name and description to the SaveModal
