+++ src/components/SaveModal/index.tsx
@@ -3,12 +3,20 @@ import { Modal, Input } from 'antd';
 import FormItem from 'antd/lib/form/FormItem';
 import { FormattedMessage } from 'umi-plugin-react/locale';
 
-export default function SaveModal(props: any) {
+interface SaveModalProps {
+  save: (name: string, description?: string) => Promise<void>;
+  setVisible: (isVisible: boolean) => void;
+  visible: boolean;
+}
+
+export default function SaveModal(props: SaveModalProps) {
   const [isLoading, setLoading] = useState(false);
+  const [name, setName] = useState('');
+  const [description, setDescription] = useState('');
 
   const handleOk = async () => {
     setLoading(true);
-    await props.save();
+    await props.save(name, description);
     setTimeout(() => {
       setLoading(false);
       props.setVisible(false);
@@ -27,17 +35,13 @@ export default function SaveModal(props: any) {
       onCancel={handleCancel}
     >
       <FormItem label={<FormattedMessage id="misc.name" defaultMessage="Name" />}>
-        <Input
-          placeholder="Name"
-          value={props.name}
-          onChange={e => props.setName(e.target.value)}
-        />
+        <Input placeholder="Name" value={name} onChange={e => setName(e.target.value)} />
       </FormItem>
       <FormItem label={<FormattedMessage id="misc.description" defaultMessage="Description" />}>
         <Input
           placeholder="Description"
-          value={props.description}
-          onChange={e => props.setDescription(e.target.value)}
+          value={description}
+          onChange={e => setDescription(e.target.value)}
         />
       </FormItem>
     </Modal>
